refactor: extract security.go, quitWithCancel helper, and selector dispatch interface
Phase 1 (read-only analysis) → docs/architecture.md

Phase 2 problem table identified 12 issues across 3 severity levels.

Phase 3 — P0/P1 fixes implemented:

P0 — handler_input.go: extract quitWithCancel()
  The cancel-stream + cancel-tool + fireSessionEnd + tea.Quit sequence
  appeared verbatim in CtrlC, CtrlD, and "exit" handlers; the "exit"
  branch was also missing m.tool.Cancel(). Single helper eliminates the
  duplication and the latent bug.

P1 — handler_input.go: selector dispatch via selectorDispatcher interface
  delegateToActiveModal had 8 identical `if m.X.Selector.IsActive()`
  blocks. Introduced selectorDispatcher interface + selectorDispatchers()
  slice method; dispatch collapses to a for-range loop. New selectors
  need only one line added to the slice.

P1 — config/permission.go → config/security.go
  Extracted 352-line security-analysis section (destructive-command
  patterns, sensitive-path checks, bash obfuscation detection,
  DenialTracking) into a dedicated file. permission.go drops from
  894 → 542 lines; security.go is 358 lines of cohesive content.

All 36 tests pass; external API unchanged.
